Refactor AdminSidebar and Student Management components; streamline sidebar items, remove unnecessary sub-items, and enhance student management functionality with improved form handling and search capabilities.
