--- src/app/admin/users/staff/page.tsx
@@ -1,428 +1,0 @@
-"use client";
-
-import { useState, useEffect, useCallback } from "react";
-import { motion } from "framer-motion";
-import { Plus, Search, Edit, Trash2 } from "lucide-react";
-import { toast } from "sonner";
-
-interface Staff {
-  _id: string;
-  name: string;
-  email: string;
-  department: string;
-  role: string;
-  phone: string;
-  createdAt: string;
-}
-
-export default function StaffManagement() {
-  const [staff, setStaff] = useState<Staff[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [search, setSearch] = useState("");
-  const [department, setDepartment] = useState("");
-  const [role, setRole] = useState("");
-  const [page, setPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(1);
-  const [isFormOpen, setIsFormOpen] = useState(false);
-  const [editingStaff, setEditingStaff] = useState<Staff | null>(null);
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    password: "",
-    department: "",
-    role: "",
-    phone: "",
-  });
-
-  const departments = [
-    "CSE",
-    "ECE",
-    "ME",
-    "CE",
-    "IT",
-    "EE",
-    "Admin",
-    "Library",
-    "Hostel",
-  ];
-  const roles = [
-    "Clerk",
-    "Lab Assistant",
-    "Office Assistant",
-    "Librarian",
-    "Hostel Warden",
-    "Security Guard",
-  ];
-
-  const fetchStaff = useCallback(async () => {
-    try {
-      const params = new URLSearchParams({
-        page: page.toString(),
-        ...(department && { department }),
-        ...(role && { role }),
-        ...(search && { search }),
-      });
-
-      const response = await fetch(`/api/user-mgmt/staff?${params}`);
-      if (!response.ok) throw new Error("Failed to fetch staff");
-
-      const data = await response.json();
-      setStaff(data.staff);
-      setTotalPages(data.totalPages);
-    } catch (error) {
-      console.log(error);
-      toast.error("Failed to fetch staff");
-    } finally {
-      setIsLoading(false);
-    }
-  }, [page, department, role, search]);
-
-  useEffect(() => {
-    fetchStaff();
-  }, [page, department, role, search, fetchStaff]);
-  
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-
-    try {
-      const url = editingStaff
-        ? `/api/user-mgmt/staff?id=${editingStaff._id}`
-        : "/api/user-mgmt/staff";
-      const method = editingStaff ? "PUT" : "POST";
-
-      const response = await fetch(url, {
-        method,
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(formData),
-      });
-
-      if (!response.ok) {
-        const error = await response.text();
-        throw new Error(error);
-      }
-
-      toast.success(
-        editingStaff
-          ? "Staff updated successfully"
-          : "Staff created successfully"
-      );
-      setIsFormOpen(false);
-      setEditingStaff(null);
-      setFormData({
-        name: "",
-        email: "",
-        password: "",
-        department: "",
-        role: "",
-        phone: "",
-      });
-      fetchStaff();
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Something went wrong"
-      );
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleDelete = async (id: string) => {
-    if (!confirm("Are you sure you want to delete this staff member?")) return;
-
-    try {
-      const response = await fetch(`/api/user-mgmt/staff?id=${id}`, {
-        method: "DELETE",
-      });
-
-      if (!response.ok) {
-        const error = await response.text();
-        throw new Error(error);
-      }
-
-      toast.success("Staff deleted successfully");
-      fetchStaff();
-    } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Failed to delete staff"
-      );
-    }
-  };
-
-  const handleEdit = (staffMember: Staff) => {
-    setEditingStaff(staffMember);
-    setFormData({
-      name: staffMember.name,
-      email: staffMember.email,
-      password: "", // Password is not included in the response
-      department: staffMember.department,
-      role: staffMember.role,
-      phone: staffMember.phone,
-    });
-    setIsFormOpen(true);
-  };
-
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <h1 className="text-2xl font-bold">Staff Management</h1>
-        <motion.button
-          whileHover={{ scale: 1.02 }}
-          whileTap={{ scale: 0.98 }}
-          onClick={() => setIsFormOpen(true)}
-          className="bg-indigo-600 hover:bg-indigo-500 text-white px-4 py-2 rounded-lg flex items-center gap-2"
-        >
-          <Plus className="w-5 h-5" />
-          Add Staff
-        </motion.button>
-      </div>
-
-      <div className="flex gap-4">
-        <div className="flex-1 relative">
-          <Search className="absolute left-3 top-1/2 -translate-y-1/2 text-gray-400 w-5 h-5" />
-          <input
-            type="text"
-            placeholder="Search by name or email..."
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            className="w-full pl-10 pr-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-          />
-        </div>
-        <select
-          value={department}
-          onChange={(e) => setDepartment(e.target.value)}
-          className="px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-        >
-          <option value="">All Departments</option>
-          {departments.map((dept) => (
-            <option key={dept} value={dept}>
-              {dept}
-            </option>
-          ))}
-        </select>
-        <select
-          value={role}
-          onChange={(e) => setRole(e.target.value)}
-          className="px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-        >
-          <option value="">All Roles</option>
-          {roles.map((r) => (
-            <option key={r} value={r}>
-              {r}
-            </option>
-          ))}
-        </select>
-      </div>
-
-      <div className="bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg overflow-hidden">
-        <table className="w-full">
-          <thead>
-            <tr className="border-b border-white/10">
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Name
-              </th>
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Email
-              </th>
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Department
-              </th>
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Role
-              </th>
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Phone
-              </th>
-              <th className="px-6 py-3 text-left text-sm font-medium text-gray-300">
-                Actions
-              </th>
-            </tr>
-          </thead>
-          <tbody className="divide-y divide-white/10">
-            {staff.map((staffMember) => (
-              <tr key={staffMember._id} className="hover:bg-white/5">
-                <td className="px-6 py-4">{staffMember.name}</td>
-                <td className="px-6 py-4">{staffMember.email}</td>
-                <td className="px-6 py-4">{staffMember.department}</td>
-                <td className="px-6 py-4">{staffMember.role}</td>
-                <td className="px-6 py-4">{staffMember.phone}</td>
-                <td className="px-6 py-4">
-                  <div className="flex gap-2">
-                    <button
-                      onClick={() => handleEdit(staffMember)}
-                      className="p-1 hover:bg-white/10 rounded-lg transition-colors"
-                    >
-                      <Edit className="w-4 h-4 text-indigo-400" />
-                    </button>
-                    <button
-                      onClick={() => handleDelete(staffMember._id)}
-                      className="p-1 hover:bg-white/10 rounded-lg transition-colors"
-                    >
-                      <Trash2 className="w-4 h-4 text-red-400" />
-                    </button>
-                  </div>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-
-      <div className="flex justify-between items-center">
-        <div className="flex gap-2">
-          <button
-            onClick={() => setPage((p) => Math.max(1, p - 1))}
-            disabled={page === 1}
-            className="px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg disabled:opacity-50"
-          >
-            Previous
-          </button>
-          <button
-            onClick={() => setPage((p) => Math.min(totalPages, p + 1))}
-            disabled={page === totalPages}
-            className="px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg disabled:opacity-50"
-          >
-            Next
-          </button>
-        </div>
-        <span className="text-sm text-gray-400">
-          Page {page} of {totalPages}
-        </span>
-      </div>
-
-      {isFormOpen && (
-        <div className="fixed inset-0 bg-black/50 backdrop-blur-sm flex items-center justify-center">
-          <motion.div
-            initial={{ opacity: 0, scale: 0.95 }}
-            animate={{ opacity: 1, scale: 1 }}
-            className="bg-gray-900 p-6 rounded-lg w-full max-w-md space-y-4"
-          >
-            <h2 className="text-xl font-bold">
-              {editingStaff ? "Edit Staff" : "Add New Staff"}
-            </h2>
-            <form onSubmit={handleSubmit} className="space-y-4">
-              <div>
-                <label className="block text-sm font-medium mb-1">Name</label>
-                <input
-                  type="text"
-                  required
-                  value={formData.name}
-                  onChange={(e) =>
-                    setFormData({ ...formData, name: e.target.value })
-                  }
-                  className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                />
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Email</label>
-                <input
-                  type="email"
-                  required
-                  value={formData.email}
-                  onChange={(e) =>
-                    setFormData({ ...formData, email: e.target.value })
-                  }
-                  className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                />
-              </div>
-              {!editingStaff && (
-                <div>
-                  <label className="block text-sm font-medium mb-1">
-                    Password
-                  </label>
-                  <input
-                    type="password"
-                    required
-                    value={formData.password}
-                    onChange={(e) =>
-                      setFormData({ ...formData, password: e.target.value })
-                    }
-                    className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                  />
-                </div>
-              )}
-              <div>
-                <label className="block text-sm font-medium mb-1">
-                  Department
-                </label>
-                <select
-                  required
-                  value={formData.department}
-                  onChange={(e) =>
-                    setFormData({ ...formData, department: e.target.value })
-                  }
-                  className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                >
-                  <option value="">Select Department</option>
-                  {departments.map((dept) => (
-                    <option key={dept} value={dept}>
-                      {dept}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Role</label>
-                <select
-                  required
-                  value={formData.role}
-                  onChange={(e) =>
-                    setFormData({ ...formData, role: e.target.value })
-                  }
-                  className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                >
-                  <option value="">Select Role</option>
-                  {roles.map((r) => (
-                    <option key={r} value={r}>
-                      {r}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              <div>
-                <label className="block text-sm font-medium mb-1">Phone</label>
-                <input
-                  type="tel"
-                  required
-                  value={formData.phone}
-                  onChange={(e) =>
-                    setFormData({ ...formData, phone: e.target.value })
-                  }
-                  className="w-full px-4 py-2 bg-white/5 backdrop-blur-sm border border-white/10 rounded-lg focus:outline-none focus:ring-2 focus:ring-indigo-500/50"
-                />
-              </div>
-              <div className="flex gap-4">
-                <button
-                  type="submit"
-                  disabled={isLoading}
-                  className="flex-1 bg-indigo-600 hover:bg-indigo-500 text-white px-4 py-2 rounded-lg disabled:opacity-50"
-                >
-                  {isLoading ? "Saving..." : "Save"}
-                </button>
-                <button
-                  type="button"
-                  onClick={() => {
-                    setIsFormOpen(false);
-                    setEditingStaff(null);
-                    setFormData({
-                      name: "",
-                      email: "",
-                      password: "",
-                      department: "",
-                      role: "",
-                      phone: "",
-                    });
-                  }}
-                  className="flex-1 bg-white/5 hover:bg-white/10 text-white px-4 py-2 rounded-lg"
-                >
-                  Cancel
-                </button>
-              </div>
-            </form>
-          </motion.div>
-        </div>
-      )}
-    </div>
-  );
-}
