+++ src/components/ui/BackgroundGridPattern.tsx
@@ -1,7 +1,7 @@
 export default function BackgroundGridPattern() {
   return (
     <div
-      className="fixed inset-0 pointer-events-none"
+      className="fixed inset-0 pointer-events-none -z-10"
       style={{
         backgroundImage: `
           linear-gradient(to right, rgba(255, 255, 255, 0.07) 1px, transparent 1px),
